+++ front/app/containers/Admin/flexible-pages/messages.ts
@@ -1,0 +1,8 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  pageHeader: {
+    id: 'app.containers.admin.flexible-pages.pageHeader',
+    defaultMessage: 'Pages',
+  },
+});