+++ front/app/typings.ts
@@ -72,7 +72,6 @@ export interface InsertConfigurationOptions<T extends { name: string }> {
   insertAfterName?: string;
   insertBeforeName?: string;
   reinsertAfterUpdate?: boolean;
-  removeName?: string;
 }
 
 export interface ILinks {